requirements.txt
autopep8
azure-kusto-data
azure-kusto-ingest
labml-nn
pandas